Move auth-ui into the login route
